fix: 경로 수정, 콘솔 한글 깨짐 오류 수정, 콤보 박스 라디오 박스 구현

+++ src/main/java/egovframework/investment/service/InvestmentDTO.java
@@ -16,7 +16,7 @@ public class InvestmentDTO implements Serializable {
     private String exchange;
     private String buyDate;
     private String memo;
-    private int memberId;
+    private Long memberId;
     
     private int pageIndex = 1;
     
@@ -67,7 +67,7 @@ public class InvestmentDTO implements Serializable {
 	public int getPageIndex() {	return pageIndex; }
 	public void setPageIndex(int pageIndex) { this.pageIndex = pageIndex; }
 	
-	public int getMemberId() { return memberId; }
-	public void setMemberId(int memberId) { this.memberId = memberId; }
+	public Long getMemberId() { return memberId; }
+	public void setMemberId(Long memberId) { this.memberId = memberId; }
 	
 }
